+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 9.3.1 - 2023-05-12
+
+## Changed
+
+* Previously, when help was requested on a subcommand method,
+  the output was of listing all subcommand methods.
+  Now, the output is now the help for that specific method.
+
 ## 9.2.0 - 2023-03-01
 
 ## Changed